11/14

+++ src/bullettest/Machine.java
@@ -1,10 +1,10 @@
 package bullettest;
 
 public class Machine {
-    public float distance;
+    public double distance;
     public String ip;
     
-    public Machine(String ip, float distance){
+    public Machine(String ip, double distance){
         this.ip = ip;
         this.distance = distance;
     }
